order-service/requirements.txt
fastapi
uvicorn
google-cloud-pubsub
python-dotenv